Revert "add new main window and windows for adding cars/customers"
This reverts commit 090cdab40ccbcd8984e81726eca124246fbaf3fa.

+++ CarRentalDatabase/dbo/Tables/Customer.sql
@@ -4,6 +4,6 @@
     [firstname] VARCHAR(50) NOT NULL, 
     [lastname] VARCHAR(50) NOT NULL, 
     [title] VARCHAR(50) NULL DEFAULT 'Herr/Frau', 
-    [age] INT NOT NULL, 
+    [age] SMALLINT NOT NULL, 
     [notes] VARCHAR(50) NULL DEFAULT 'N/A'
 )
